move to own class
